.ipynb_checkpoints/requirements-checkpoint.txt
streamlit
tflite-runtime
imageio[ffmpeg]
Pillow
numpy
